More plumbing in the eval pipeline
