+++ packages/core/src/hooks/index.ts
@@ -1,3 +1,4 @@
+import type { ScorerHookData } from '../eval';
 import type { Metric, MetricResult } from '../eval/metric';
 import type { TestInfo } from '../eval/types';
 
@@ -31,24 +32,6 @@ type GenerationHookData = {
   runId: string;
   agentName: string;
   instructions: string;
-};
-
-type ScoringSource = 'LIVE';
-type ScoringEntityType = 'AGENT';
-
-export type ScorerHookData = {
-  runId: string;
-  traceId?: string;
-  scorer: Record<string, any>;
-  input: string;
-  output: string;
-  additionalContext?: Record<string, any>;
-  resourceId?: string;
-  threadId?: string;
-  source: ScoringSource;
-  entity: Record<string, any>;
-  entityType: ScoringEntityType;
-  runtimeContext: Record<string, any>;
 };
 
 export function registerHook(hook: AvailableHooks.ON_EVALUATION, action: Handler<EvaluationHookData>): void;
